--- evora/__init__.py
@@ -1,5 +1,0 @@
-
-try:
-    from .andor import *
-except ImportError:
-    from .dummy import *
