+++ evora/andor_error.py
@@ -88,3 +88,13 @@ ERROR_CODES = {
     20991: ["DRV_NOT_SUPPORTED", "None"],
     20992: ["DRV_NOT_AVAILABLE", "None"]
 }
+
+class AndorCameraError(Exception):
+    def __init__(self, error_code):
+        self.error_code = error_code
+        super().__init__(ERROR_CODES[self.error_code])
+
+    def __str__(self):
+        error_info = ERROR_CODES[self.error_code]
+        return f'({self.error_code}) {error_info[0]}. {error_info[1]}'
+
