+++ evora/pyproject.toml
@@ -1,0 +1,2 @@
+[tool.setuptools]
+packages = ["andor_routines", "andor", "dummy", "andor_error"]
