--- pyproject.toml
@@ -1,17 +1,0 @@
-[tool.poetry]
-name = "instruments"
-version = "0.1.0"
-description = ""
-authors = ["Your Name <you@example.com>"]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = "^3.9"
-flask = "^2.0.1"
-
-[tool.poetry.dev-dependencies]
-isort = "^5.9.3"
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
